[sparkle][front] Trim the model picker highlight, add a replay switch
Fold the hook into its only consumer and cut the commentary back to the traps
that would actually bite someone editing this: why the listener is on
`pointerdown`, why the swept layers rest off-screen, why the host stays mounted,
and why the dismissal count is ref-guarded. The rest was restating the diff.

281 insertions down to 218, and one file fewer.

Also add `?replayHighlight`, which forces the highlight on and spends nothing.
Reviewing a deploy preview otherwise means clearing localStorage by hand once the
two dismissals are gone.
